Move functions for HuggingFaceHub interactions to own file
