+++ tests/test_cli.py
@@ -18,6 +18,7 @@ from huggingface_hub.utils import HFValidationError
 
 import annif.cli
 import annif.cli_util
+import annif.hfh_util
 import annif.parallel
 
 runner = CliRunner(env={"ANNIF_CONFIG": "annif.default_config.TestingConfig"})
@@ -1140,7 +1141,7 @@ def test_archive_dir(testdatadir):
     open(os.path.join(str(dirpath), "foo.txt"), "a").close()
     open(os.path.join(str(dirpath), "-train.txt"), "a").close()
 
-    fobj = annif.cli_util._archive_dir(dirpath)
+    fobj = annif.hfh_util._archive_dir(dirpath)
     assert isinstance(fobj, io.BufferedRandom)
 
     with zipfile.ZipFile(fobj, mode="r") as zfile:
@@ -1150,7 +1151,7 @@ def test_archive_dir(testdatadir):
 
 
 def test_get_project_config(app_project):
-    result = annif.cli_util._get_project_config(app_project)
+    result = annif.hfh_util._get_project_config(app_project)
     assert isinstance(result, io.BytesIO)
     string_result = result.read().decode("UTF-8")
     assert "[dummy-en]" in string_result
@@ -1175,7 +1176,7 @@ def hf_hub_download_mock_side_effect(filename, repo_id, token, revision):
     "huggingface_hub.hf_hub_download",
     side_effect=hf_hub_download_mock_side_effect,
 )
-@mock.patch("annif.cli_util.copy_project_config")
+@mock.patch("annif.hfh_util.copy_project_config")
 def test_download_dummy_fi(
     copy_project_config, hf_hub_download, list_repo_files, testdatadir
 ):
@@ -1233,7 +1234,7 @@ def test_download_dummy_fi(
     "huggingface_hub.hf_hub_download",
     side_effect=hf_hub_download_mock_side_effect,
 )
-@mock.patch("annif.cli_util.copy_project_config")
+@mock.patch("annif.hfh_util.copy_project_config")
 def test_download_dummy_fi_and_en(
     copy_project_config, hf_hub_download, list_repo_files, testdatadir
 ):
@@ -1352,7 +1353,7 @@ def test_download_hf_hub_download_failed(
 def test_unzip_archive_initial(testdatadir):
     dirpath = os.path.join(str(testdatadir), "projects", "dummy-fi")
     fpath = os.path.join(str(dirpath), "file.txt")
-    annif.cli_util.unzip_archive(
+    annif.hfh_util.unzip_archive(
         os.path.join("tests", "huggingface-cache", "projects", "dummy-fi.zip"),
         force=False,
     )
@@ -1371,7 +1372,7 @@ def test_unzip_archive_no_overwrite(testdatadir):
     with open(fpath, "wt") as pf:
         print("Existing content", file=pf)
 
-    annif.cli_util.unzip_archive(
+    annif.hfh_util.unzip_archive(
         os.path.join("tests", "huggingface-cache", "projects", "dummy-fi.zip"),
         force=False,
     )
@@ -1387,7 +1388,7 @@ def test_unzip_archive_overwrite(testdatadir):
     with open(fpath, "wt") as pf:
         print("Existing content", file=pf)
 
-    annif.cli_util.unzip_archive(
+    annif.hfh_util.unzip_archive(
         os.path.join("tests", "huggingface-cache", "projects", "dummy-fi.zip"),
         force=True,
     )
@@ -1400,10 +1401,10 @@ def test_unzip_archive_overwrite(testdatadir):
 
 
 @mock.patch("os.path.exists", return_value=True)
-@mock.patch("annif.cli_util._compute_crc32", return_value=0)
+@mock.patch("annif.hfh_util._compute_crc32", return_value=0)
 @mock.patch("shutil.copy")
 def test_copy_project_config_no_overwrite(copy, _compute_crc32, exists):
-    annif.cli_util.copy_project_config(
+    annif.hfh_util.copy_project_config(
         os.path.join("tests", "huggingface-cache", "dummy-fi.cfg"), force=False
     )
     assert not copy.called
@@ -1412,7 +1413,7 @@ def test_copy_project_config_no_overwrite(copy, _compute_crc32, exists):
 @mock.patch("os.path.exists", return_value=True)
 @mock.patch("shutil.copy")
 def test_copy_project_config_overwrite(copy, exists):
-    annif.cli_util.copy_project_config(
+    annif.hfh_util.copy_project_config(
         os.path.join("tests", "huggingface-cache", "dummy-fi.cfg"), force=True
     )
     assert copy.called
